refactor: moved modules command deleted old
