+++ cognite_toolkit/_cdf_tk/apps/__init__.py
@@ -1,0 +1,4 @@
+from ._landing_app import LandingApp
+from ._modules_app import ModulesApp
+
+__all__ = ["LandingApp", "ModulesApp"]
