+++ cognite_toolkit/_cdf_tk/prototypes/commands/__init__.py
@@ -1,5 +1,4 @@
 from .dump_assets import DumpAssetsCommand
 from .dump_timeseries import DumpTimeSeriesCommand
-from .modules import ModulesCommand
 
-__all__ = ["ModulesCommand", "DumpAssetsCommand", "DumpTimeSeriesCommand"]
+__all__ = ["DumpAssetsCommand", "DumpTimeSeriesCommand"]
