--- cognite_toolkit/custom_modules/README.md
@@ -1,3 +1,0 @@
-# local_modules directory
-
-See the [documentation](https://docs.cognite.com/cdf/deploy/cdf_toolkit/guides/build_modules) for an introduction.
